play_random_game(): Do not give first move special significance
Previously, first move was reported back to the caller for the use of the
Montecarlo engine; also, there was special policy hinting for the first move,
and an explicit suicide check, which is actually harmful for UCT (if suicides
are only remaining moves on the board, the playout always returns error and the
nodes are never expanded to handle the suicides properly).

Now, Montecarlo is playing out the first move by itself; play_random_game() was
considerably simplified.