UCT: Reuse measurements from same board positions
If two board positions with only different order of play but same content are
reached, the playout statistics and continuations are reused for all the tree
nodes. This gets very hairy when removing nodes, and we don't quite get that
right now, but this shouldn't hurt us much hopefully. The reuse rate seems to
be around 10-20%.