UCT: Remove sharepos support
This horribly complicated all the code, was buggy, and could
never work anyway since e.g. captures count was not hashed.
We might try it again in the future but so far it is too bug-ridden,
and complicates the code too much; besides, it seems basically useless
on large boards and only reuses small amount of positions on 9x9.