Move time_prep() from uct/uct.c to time_stop_conditions() of timeinfo.*
This way, all the timeinfo setup/decisions code is at a single place and
abstracted out of a particular engine.

I think that this way, we might do also some cleanups not to forcibly
convert TT_TOTAL to TT_MOVE but instead do the conversion only in
time_stop_conditions(). But that's not high-priority.