BOARD_TRAITS support for incremental per-intersection information
Board traits enable storage of per-point information about its current
tactical features, maintained incrementally. Currently, only the capture
trait is maintained, but in the future, at least atari and selfatari
information should be maintained as well.

This of course slows the regular users down, which is also a reason
why the code is disabled by default. However, it is an essential
groundwork for fully-incremental probability distribution support.