Remove the General Pattern Matcher
This has been a costly failed experiment. I spent many weeks of hard
work trying to make it work, unfortunately it seems unlikely that it
will ever attain the required performance. Therefore, let's remove it
from the current version so that it does not clutter up the source.
It is still available in the history and I expect to yet dig bits out
of there.

This leaves the board traits support orphaned, and for future probdist
experiments we will need to re-add the board gamma support. But I have
not abandoned probability distribution support entirely; in the moggypd
branch I'm trying a different approach, using direct Moggy policy hints
for pattern matching, instead of building up an independent matcher.