board_large(): Introduce, use instead of current 9-or-19 tests
Instead, new threshold is 15x15. It is just my feeling that from this
point on, the board behaves more like 19x19. 13x13 is kind of
borderline, but I do not think anyone really cares much about this
board size anyway; and I do like to run Pachi on 15x15 occassionally.