Remove the treepool feature
The treepool has been a failed experiment so far, and it does not seem
likely we will get to make it a real playing strength contribution
anytime soon. Therefore, let's leave it just in the git history.

I have kept the 'significant node' feature since it is used also in some
unmerged branches, is a small piece of code and can come useful for
other stuff I have in mind.