UCB1AMAF: vloss_sqrt to re-normalize virtual loss for better many-threads performance
Rescale virtual loss value to square root of #threads. This mitigates
the number of virtual losses added in case of a large amount of
threads; it seems that with linear virtual losses, overly diverse
exploration caused by this may cause a wrong mean value computed
for the parent node.

Off by default for now.