selfatari_cousin(): Suggest counter-captures
When looking for alternatives of a self-atari move, aside of the
other liberty also look if we can counter-capture an opponent's
group.

This improves handling of some false-eyeish corner situations,
where we will more often capture the falsifying opponent stones
rather than connecting the pseudo-false eye (which can be fatal
to the status of the group).