Moggy selfatarirate: 95 -> 60
This is a drastic change that I didn't want to make, and this is also
something where I left off Pachi development the last time in Aug 2014.
It turns out that this change is +50 Elo against GNUGo on 15x15:

500s S.D., -50.5 komi
games elo-sd  elo     elo+sd
1823  420     428     437     selfatarirate=10
2030  423     431     439     selfatarirate=25
1873  437     445     453     selfatarirate=50
1956  428     436     444     selfatarirate=75
1087  412     423     433     selfatarirate=80
1991  390     397     405     selfatarirate=90
1935  391     398     406     selfatarirate=95

1000s S.D., -65.5 komi
games elo-sd  elo     elo+sd
2587  336     343     350     selfatarirate=1
2702  342     349     356     selfatarirate=20
2604  350     357     364     selfatarirate=40
2729  349     356     363     selfatarirate=60
2675  320     327     334     selfatarirate=90
2711  301     307     314     selfatarirate=95

2000s S.D., -75.5 komi (this is already very high, distorts results?)
games elo-sd  elo     elo+sd
2892  262     269     276     selfatarirate=1
2961  258     265     272     selfatarirate=60
2987  229     236     243     selfatarirate=95

I didn't want to just commit this parameter change but wanted to check
about its cause, but I don't know if I'll ever get around to that, so
let's just grab this change for now so that we have a clean plate here.