Travis CI is failing
I noticed that the CI is failing for PR #41. It is because the installation of clang Ubuntu package is fetching a file from llvm.org and that file is no longer there, see http://stackoverflow.com/questions/37558151/travis-ci-build-failing for a discussion.

Immediate solution would be to temporarily remove clang installation from the build. Longer term solution depends what comes out of the debate referenced at that SO.com question.