mm tool: detlef's patch
http://computer-go.org/pipermail/computer-go/2013-May/005979.html

Patch for Remi's original mm-tools

Hi,

I contacted Remi before, but he is using a much improved version of his
tools now and suggested I could share the patch here.

It speeds up the mm-tools by a factor of 50 in the iteration loop if you
use it for a high number of gammas (like in case of large patterns).

I hope it helps:)

This patch is minimalistic, so you should be able to check it's
correctness by reading. I have some more changes which allow the gcc
vectorizer to improve some of the initial loops, but this does not make
such a big difference.

Detlef