mm: fixed spatial dictionary: only use patterns from played moves
This is best as it captures interesting situations around moves we want to play.
The other way (competition, scanning all available moves on the board) was a
mistake of course, yields lots of frequently occuring but useless patterns for
making a decision. Ended up with lots of spatials with 1.0 gamma in pattterns_mm.gamma.

Retraining with new spatial dictionary prediction rate jumps from 30% to 37%.