tree cleanup: removed t->board
i guess it was there mostly for convenience at a time when
coord2sstr() etc needed a board argument.

no real reason to have it anymore as far as i can tell,
anything that uses it currently does fine without it,
plus i'd rather make it very explicit what board we're
working on given that it can create threads issues.