josekifix engine: decouple josekifix logic from gtp layer
if josekifix is enabled josekifix_engine replaces uct as main engine
and acts as middle man between gtp layer and uct engine. it also takes
care of forwarding gtp notifications to external joseki engine so it
can stay in sync.

this way there is zero need to hardcode anything josekifix related in
the gtp layer, as far as it's concerned it's just another engine.

only hooks remaining are in pachi.c (uct/josekifix_engine switch)
and in uct.c genmove(). gtp layer back to clean state.